refactor(core): move pr_create_verify into the core/merge package
