+++ BLUEPRINT.md
@@ -604,7 +604,7 @@ Why this architecture exists, observed repeatedly: durability comes from **enfor
 
 12. **An outcome-claiming completion carries a resolvable artifact pointer, fail-closed.** The out-of-band completion surface (`tasks complete --note`) records externally-landed work. `teatree.core.review.completion_evidence` makes two separate judgments: a note *asserts an outcome* when an `OUTCOME_CLAIM_KINDS` verb (`merged`/`posted`/`shipped`/`deployed` + synonyms) co-occurs with a context cue (internal idioms like `merge conflict` are stripped first), and an asserting note must carry a *resolvable pointer* an auditor can follow — a URL, a cued git SHA, an MR/PR/issue ref, a note id, a Slack ts, or a real/dotted-module path — else `check_completion_evidence` refuses (`CompletionEvidenceError`). A completion with no note, or asserting no outcome, is never gated (the structural form of invariant 6's "done claims require evidence"); the exact matchers live in `completion_evidence.py`. A `failed → completed` resolution requires a mandatory `--note` ([#1949](https://github.com/souliane/teatree/issues/1949)); `claimed → completed` stays note-optional.
 
-13. **Post-write side effects are confirmed by re-reading, not assumed from the write's own response** ([#1192](https://github.com/souliane/teatree/issues/1192)). A write's own success response (`ok: true`, a harness tool call returning without error) is not proof its effect is actually visible — eventual consistency, a stale replica, or a race can make a claimed write unobservable moments later. `teatree.core.verify_by_reread.verify_by_reread` is the small core contract: given a `reread` callable that independently re-observes the target system, it returns a `RereadOutcome` (`confirmed` / `not_confirmed(reason)`) and never raises — a verification failure is a signal to log or retry, never a reason to crash the write path. Call sites: Slack `reactions.add` (`teatree.backends.slack.reactions.add_reaction_verified`, wired into `add_reactions_for_transition` / `add_approval_reaction`) re-reads via `reactions.get` before trusting the write; and `create_pr` ([#1194](https://github.com/souliane/teatree/issues/1194), `teatree.core.pr_create_verify.verify_pr_exists`) re-reads a just-created PR's open-state on the ship / `ensure-pr` path before recording the URL, so a phantom PR (a 404 re-read) is reported failed and never advances the FSM. Inbound-event ingestion applies the sibling shape — a failed drain records the error and retries with backoff, then dead-letters (`IncomingEvent.record_failure`, [#673](https://github.com/souliane/teatree/issues/673)) so a poison event never silently drops nor blocks the queue. Direct MCP Slack writes are refused at the `PreToolUse` boundary (`handle_block_mcp_slack_write`, [#1196](https://github.com/souliane/teatree/issues/1196), kill-switch `[teatree] mcp_slack_write_gate_enabled`) so every Slack write routes through the `t3` CLI egress chokepoint under this contract.
+13. **Post-write side effects are confirmed by re-reading, not assumed from the write's own response** ([#1192](https://github.com/souliane/teatree/issues/1192)). A write's own success response (`ok: true`, a harness tool call returning without error) is not proof its effect is actually visible — eventual consistency, a stale replica, or a race can make a claimed write unobservable moments later. `teatree.core.verify_by_reread.verify_by_reread` is the small core contract: given a `reread` callable that independently re-observes the target system, it returns a `RereadOutcome` (`confirmed` / `not_confirmed(reason)`) and never raises — a verification failure is a signal to log or retry, never a reason to crash the write path. Call sites: Slack `reactions.add` (`teatree.backends.slack.reactions.add_reaction_verified`, wired into `add_reactions_for_transition` / `add_approval_reaction`) re-reads via `reactions.get` before trusting the write; and `create_pr` ([#1194](https://github.com/souliane/teatree/issues/1194), `teatree.core.merge.pr_create_verify.verify_pr_exists`) re-reads a just-created PR's open-state on the ship / `ensure-pr` path before recording the URL, so a phantom PR (a 404 re-read) is reported failed and never advances the FSM. Inbound-event ingestion applies the sibling shape — a failed drain records the error and retries with backoff, then dead-letters (`IncomingEvent.record_failure`, [#673](https://github.com/souliane/teatree/issues/673)) so a poison event never silently drops nor blocks the queue. Direct MCP Slack writes are refused at the `PreToolUse` boundary (`handle_block_mcp_slack_write`, [#1196](https://github.com/souliane/teatree/issues/1196), kill-switch `[teatree] mcp_slack_write_gate_enabled`) so every Slack write routes through the `t3` CLI egress chokepoint under this contract.
 
 ### 17.2 The flywheel — 17.8 Orchestrator-as-keystone contract
