+++ src/teatree/core/verify_by_reread.py
@@ -19,7 +19,7 @@ Call sites wire this in (#1192, subsuming #1193 and #1202; #1194):
 
 *   :func:`teatree.backends.slack.reactions.add_reaction_verified` re-reads a
     posted Slack reaction via ``reactions.get`` before trusting ``reactions.add``.
-*   :func:`teatree.core.pr_create_verify.verify_pr_exists` re-reads a just-created
+*   :func:`teatree.core.merge.pr_create_verify.verify_pr_exists` re-reads a just-created
     PR's open-state before the ship/ensure path trusts ``create_pr`` and records
     the URL, so a phantom PR (a 404 re-read) never advances the FSM (#1194).
